examples/ml-pipeline/requirements.txt
# Core ML libraries
tensorflow==2.13.0
pytorch-lightning==2.0.6
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Data processing
pyspark==3.4.1
dask==2023.7.1
great-expectations==0.17.1
feast==0.30.1
petastorm==0.12.0

# MLOps
mlflow==2.6.0
optuna==3.3.0
wandb==0.15.8
evidently==0.4.0
deepchecks==0.17.0

# Orchestration
apache-airflow==2.6.3
prefect==2.11.3

# Serving
fastapi==0.103.1
uvicorn==0.23.2
grpcio==1.57.0
grpcio-tools==1.57.0
triton-client==2.36.0

# Monitoring
prometheus-client==0.17.1
grafana-dashboard-builder==0.6.0

# Development
pytest==7.4.0
black==23.7.0
pylint==2.17.5
mypy==1.5.1
jupyterlab==4.0.5
jupyter==1.0.0

# Documentation
sphinx==7.2.5
sphinx-rtd-theme==1.3.0

# Containerization
docker==6.1.3
kubernetes==27.2.0 